Exec wget with default user when using cache_dir
When using the cache_dir attribute in wget::fetch we don't need to exec
wget with the execuser because only the destination file has to be owned
by it. In fact, in this case it's better to exec wget with the default
user (usually 'root') because the cache_dir may not be writeable by the
execuser.