added md5 $source_hash support to Wget::Fetch
If a $source_hash is given any existing file must match it or it
is removed prior a triggered redownload.
If a $source_hash is given the downloaded file must match it.

The system test uses an RFC to test the md5 sum check, this
can fail even without a failing test, if the RFC document
is changed.